requirements.txt
aiogram==3.6.0
aiohttp
python-dotenv
yt-dlp
